requirements.txt
# Core dependencies
pyyaml>=6.0
python-dateutil>=2.8.2
requests>=2.31.0
python-dotenv>=1.0.0

# CLI and UI
colorama>=0.4.6
rich>=13.0.0

# Validation
validators>=0.20.0
dnspython>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
